dieselrecommerce/requirements.txt
backports.csv==1.0.7
blessed==1.15.0
cached-property==1.5.1
cement==2.8.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.3.9
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.5
django-import-export==1.2.0
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.0.1
future==0.16.0
idna==2.7
jdcal==1.4.1
jsonschema==2.6.0
numpy==1.17.2
odfpy==1.4.0
openpyxl==2.6.3
pandas==0.25.1
pathspec==0.5.9
python-dateutil==2.8.0
pytz==2019.2
PyYAML==3.13
requests==2.22.0
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
tablib==0.13.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.25.6
websocket-client==0.56.0
xlrd==1.2.0
xlwt==1.3.0
